+++ src/main/java/com/venky/swf/plugins/calendar/db/model/SpecialWorkDay.java
@@ -11,8 +11,8 @@ public interface SpecialWorkDay extends Model{
 	@UNIQUE_KEY
 	@IS_NULLABLE(false)
 	@HIDDEN
-	public Integer getWorkCalendarId();
-	public void setWorkCalendarId(Integer id);
+	public Long getWorkCalendarId();
+	public void setWorkCalendarId(Long id);
 	public WorkCalendar getWorkCalendar();
 	
 	@UNIQUE_KEY
@@ -23,7 +23,7 @@ public interface SpecialWorkDay extends Model{
 	public void setPurpose(String Purpose);
 
 	@UNIQUE_KEY
-	public Integer getWorkSlotId();
-	public void setWorkSlotId(Integer id);
+	public Long getWorkSlotId();
+	public void setWorkSlotId(Long id);
 	public WorkSlot getWorkSlot();
 }
